Match and build replacements for <img> tags in content
Adds a Data Migration class, which will hold all the helper methods for
converting old <img> tags to the shortcodes for this plugin.

Initial methods are just a regex matcher to find images in a post
content, and a callback which can take the attributes matched by the
regex and build a shortcode from them.