Adds test for media_send_to_editor functionality
In order for the test to pass, I had to move the filter to a higher
priority to prevent it being clobbered by the filter that applies the
`[caption]` shortcode. I think that's only necessary in the test suite
environment where the plugin is loaded before init... but it doesn't
hurt anywhere else, so I'm leaving it in.