Abstract out function to search string for tags to replace
There are going to be cases where you want to search strings other than
post content, so the functions that search for <img> tags and [caption]
shortcodes to replace should be able to operate on strings, not just
post objects.

Tests are still the same, they're just moved up to the higher level
function.