Actually this needs to be rawurlencode
My mistake in #68 - I wasn't understanding the difference between
rawurlencode and urlencode, and as a result I didn't catch the mismatch
until writing my integration tests elsewhere. This replaces the
urlencode filter on encoded fields sent to the editor with
`rawurlencode()` which matches the way these fields are decoded on
display.